GCW-3432 Move expectation fn to test utils
